+++ mne/tests/test_rank.py
@@ -72,11 +72,12 @@ def test_rank_estimation():
 
 
 @pytest.mark.slowtest
+@pytest.mark.parametrize('meg', ('separate', 'combined'))
 @pytest.mark.parametrize('rank_method, proj', [('info', True),
                                                ('info', False),
                                                (None, True),
                                                (None, False)])
-def test_cov_rank_estimation(rank_method, proj):
+def test_cov_rank_estimation(rank_method, proj, meg):
     """Test cov rank estimation."""
     # Test that our rank estimation works properly on a simple case
     evoked = read_evokeds(ave_fname, condition=0, baseline=(None, 0),
@@ -94,15 +95,13 @@ def test_cov_rank_estimation(rank_method, proj):
 
     raw_sss = read_raw_fif(hp_fif_fname)
     assert not _has_eeg_average_ref_proj(raw_sss.info['projs'])
-    raw_sss.add_proj(compute_proj_raw(raw_sss))
+    raw_sss.add_proj(compute_proj_raw(raw_sss, meg=meg))
 
     cov_sample = compute_raw_covariance(raw_sample)
-    cov_sample_proj = compute_raw_covariance(
-        raw_sample.copy().apply_proj())
+    cov_sample_proj = compute_raw_covariance(raw_sample.copy().apply_proj())
 
     cov_sss = compute_raw_covariance(raw_sss)
-    cov_sss_proj = compute_raw_covariance(
-        raw_sss.copy().apply_proj())
+    cov_sss_proj = compute_raw_covariance(raw_sss.copy().apply_proj())
 
     picks_all_sample = pick_types(raw_sample.info, meg=True, eeg=True)
     picks_all_sss = pick_types(raw_sss.info, meg=True, eeg=True)
@@ -157,7 +156,7 @@ def test_cov_rank_estimation(rank_method, proj):
 
             expected_rank = n_meg + n_eeg
             if rank_method is None:
-                if not has_sss:  # XXX see below and gh-5146
+                if meg == 'combined' or not has_sss:
                     if proj:
                         expected_rank -= n_projs_info
                     else:
@@ -174,18 +173,22 @@ def test_cov_rank_estimation(rank_method, proj):
 @testing.requires_testing_data
 @pytest.mark.parametrize('fname, rank_orig', ((hp_fif_fname, 120),
                                               (mf_fif_fname, 67)))
-@pytest.mark.parametrize('n_proj', (0, 10))
-def test_maxfilter_get_rank(n_proj, fname, rank_orig):
+@pytest.mark.parametrize('n_proj, meg', ((0, 'combined'),
+                                         (10, 'combined'),
+                                         (10, 'separate')))
+def test_maxfilter_get_rank(n_proj, fname, rank_orig, meg):
     """Test maxfilter rank lookup."""
     raw = read_raw_fif(fname).crop(0, 5).load_data().pick_types()
     assert raw.info['projs'] == []
     mf = raw.info['proc_history'][0]['max_info']
     assert mf['sss_info']['nfree'] == rank_orig
     assert _get_rank_sss(raw) == rank_orig
-    rank = rank_orig - 2 * n_proj
+    mult = 1 + (meg == 'separate')
+    rank = rank_orig - mult * n_proj
     if n_proj > 0:
         # Let's do some projection
-        raw.add_proj(compute_proj_raw(raw, n_mag=n_proj, n_grad=n_proj))
+        raw.add_proj(compute_proj_raw(raw, n_mag=n_proj, n_grad=n_proj,
+                                      meg=meg, verbose=True))
     raw.apply_proj()
     data_orig = raw[:][0]
 
@@ -195,8 +198,7 @@ def test_maxfilter_get_rank(n_proj, fname, rank_orig):
     with pytest.raises(TypeError, match='must be a string or a number'):
         _estimate_rank_raw(raw, tol=None)
 
-    # XXX this should be "rank" not "rank_orig", see gh-5146
-    allowed_rank = [rank_orig]
+    allowed_rank = [rank_orig if meg == 'separate' else rank]
     if fname == mf_fif_fname:
         # Here we permit a -1 because for mf_fif_fname we miss by 1, which is
         # probably acceptable. If we use the entire duration instead of 5 sec